requirements.txt
appdirs==1.4.3
cached-property==1.3.0
colorama==0.3.9
Django==1.11.1
django-cors-headers==2.0.2
djangorestframework==3.6.3
docker==2.3.0
docker-compose==1.13.0
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
jsonschema==2.6.0
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2017.2
PyYAML==3.12
requests==2.11.1
six==1.10.0
texttable==0.8.8
uWSGI==2.0.15
websocket-client==0.40.0
